LayoutTests/platform/ios/imported/w3c/web-platform-tests/html/semantics/forms/constraints/form-validation-willValidate-expected.txt


PASS [INPUT in HIDDEN status] Must be barred from the constraint validation
PASS [INPUT in BUTTON status] Must be barred from the constraint validation
PASS [INPUT in RESET status] Must be barred from the constraint validation
PASS [BUTTON in BUTTON status] Must be barred from the constraint validation
PASS [BUTTON in RESET status] Must be barred from the constraint validation
FAIL [fieldset]  Must not be barred from the constraint validation assert_true: The willValidate attribute should be true. expected true got false
FAIL [output]  Must not be barred from the constraint validation assert_true: The willValidate attribute should be true. expected true got false
PASS [object]  Must be barred from the constraint validation
PASS [keygen]  Must be barred from the constraint validation
PASS [INPUT in TEXT status] Must be barred from the constraint validation if it is disabled
PASS [INPUT in TEXT status] The willValidate attribute must be true if an element is mutable
PASS [INPUT in TEXT status] Must be barred from the constraint validation if it is readonly
PASS [INPUT in TEXT status] The willValidate attribute must be false if it has a datalist ancestor
PASS [INPUT in SEARCH status] Must be barred from the constraint validation if it is disabled
PASS [INPUT in SEARCH status] The willValidate attribute must be true if an element is mutable
PASS [INPUT in SEARCH status] Must be barred from the constraint validation if it is readonly
PASS [INPUT in SEARCH status] The willValidate attribute must be false if it has a datalist ancestor
PASS [INPUT in TEL status] Must be barred from the constraint validation if it is disabled
PASS [INPUT in TEL status] The willValidate attribute must be true if an element is mutable
PASS [INPUT in TEL status] Must be barred from the constraint validation if it is readonly
PASS [INPUT in TEL status] The willValidate attribute must be false if it has a datalist ancestor
PASS [INPUT in URL status] Must be barred from the constraint validation if it is disabled
PASS [INPUT in URL status] The willValidate attribute must be true if an element is mutable
PASS [INPUT in URL status] Must be barred from the constraint validation if it is readonly
PASS [INPUT in URL status] The willValidate attribute must be false if it has a datalist ancestor
PASS [INPUT in EMAIL status] Must be barred from the constraint validation if it is disabled
PASS [INPUT in EMAIL status] The willValidate attribute must be true if an element is mutable
PASS [INPUT in EMAIL status] Must be barred from the constraint validation if it is readonly
PASS [INPUT in EMAIL status] The willValidate attribute must be false if it has a datalist ancestor
PASS [INPUT in PASSWORD status] Must be barred from the constraint validation if it is disabled
PASS [INPUT in PASSWORD status] The willValidate attribute must be true if an element is mutable
PASS [INPUT in PASSWORD status] Must be barred from the constraint validation if it is readonly
PASS [INPUT in PASSWORD status] The willValidate attribute must be false if it has a datalist ancestor
FAIL [INPUT in DATETIME status] The datetime type must be supported. assert_equals: The datetime type should be supported. expected "datetime" but got "text"
PASS [INPUT in DATE status] Must be barred from the constraint validation if it is disabled
PASS [INPUT in DATE status] The willValidate attribute must be true if an element is mutable
PASS [INPUT in DATE status] Must be barred from the constraint validation if it is readonly
PASS [INPUT in DATE status] The willValidate attribute must be false if it has a datalist ancestor
PASS [INPUT in MONTH status] Must be barred from the constraint validation if it is disabled
PASS [INPUT in MONTH status] The willValidate attribute must be true if an element is mutable
PASS [INPUT in MONTH status] Must be barred from the constraint validation if it is readonly
PASS [INPUT in MONTH status] The willValidate attribute must be false if it has a datalist ancestor
PASS [INPUT in WEEK status] Must be barred from the constraint validation if it is disabled
PASS [INPUT in WEEK status] The willValidate attribute must be true if an element is mutable
PASS [INPUT in WEEK status] Must be barred from the constraint validation if it is readonly
PASS [INPUT in WEEK status] The willValidate attribute must be false if it has a datalist ancestor
PASS [INPUT in TIME status] Must be barred from the constraint validation if it is disabled
PASS [INPUT in TIME status] The willValidate attribute must be true if an element is mutable
PASS [INPUT in TIME status] Must be barred from the constraint validation if it is readonly
PASS [INPUT in TIME status] The willValidate attribute must be false if it has a datalist ancestor
FAIL [INPUT in COLOR status] The color type must be supported. assert_equals: The color type should be supported. expected "color" but got "text"
PASS [INPUT in FILE status] Must be barred from the constraint validation if it is disabled
PASS [INPUT in FILE status] The willValidate attribute must be true if an element is mutable
PASS [INPUT in FILE status] Must be barred from the constraint validation if it is readonly
PASS [INPUT in FILE status] The willValidate attribute must be false if it has a datalist ancestor
PASS [INPUT in SUBMIT status] Must be barred from the constraint validation if it is disabled
PASS [INPUT in SUBMIT status] The willValidate attribute must be true if an element is mutable
PASS [INPUT in SUBMIT status] Must be barred from the constraint validation if it is readonly
PASS [INPUT in SUBMIT status] The willValidate attribute must be false if it has a datalist ancestor
PASS [BUTTON in SUBMIT status] Must be barred from the constraint validation
PASS [BUTTON in SUBMIT status] The willValidate attribute must be true if an element is mutable
PASS [BUTTON in SUBMIT status] The willValidate attribute must be false if it has a datalist ancestor
PASS [select]  Must be barred from the constraint validation
PASS [select]  The willValidate attribute must be true if an element is mutable
PASS [select]  The willValidate attribute must be false if it has a datalist ancestor
PASS [textarea]  Must be barred from the constraint validation
PASS [textarea]  The willValidate attribute must be true if an element is mutable
PASS [textarea]  The willValidate attribute must be false if it has a datalist ancestor
